Simplify conditional in `make_noise`
The `make_noise` method has a bewildering array of dependent
(`if/elsif/else`) and independent (`if...end`) conditions. Some of these
get nested pretty deeply, some are inline, some are re-testing
conditions we've seen before.

Looking at all of them however, we can tell that they only ever branch
on two variables: species, and loudness. We effectively have a tuple of
`(Bird | Animal) x (Quiet | Loud)`, for only 4 possible combinations.

This can be seen visually by doing a [decision table](https://www.hillelwayne.com/decision-tables)
 analysis of inputs and outputs:

|        | Quiet               | Loud                |
|--------|---------------------|---------------------|
| Bird   | quiet bird noise    | loud bird noise     |
| Animal | single animal noise | double animal noise |

We can re-write the whole method as a flat `if/elsif/elsif/elsif` to
handle all 4 options and move the appropriate behavior to that branch.
This flat structure is much easier to read!

Now that all the distractions of the condition are gone, we can notice
that `make_animal_noise` and `make_bird_noise` seem to behave
inconsistently with each other. That's probably the next problem to
tackle.

Also, there is some **undefined behavior** for situations where the
species is neither an animal nor a bird. There are no tests for it. I
chose to return `nil` for those cases, which is what the original code
was doing.