Extract raise to a guard clause
There's a raise condition if we try to make quiet hadedah sounds. Since
this is exceptional, I felt better pulling it out to a guard clause. It
also has the benefit leaving the animal and bird branches identical to
each other.

Note that there was some untested (unwanted?) behavior in the original
implementation where the program would still output "squawk" for quiet
hadedahs before raising. The guard clause implementation not longer does
this.