Remove bird/animal checks
After the last refactor, we now see that animal/bird distinction no
longer affects the output, only loudness. Thus we can remove the
conditions on animal/bird and collapse to just outputing the sound once
or twice based on desired loudness.

Note that the undefined behavior for unknown species is now a bit
different. We will attempt to output a `nil` sound for them.